backend/requirements/base.txt
Django==5.2.12
djangorestframework==3.17.1
django-cors-headers==4.9.0
python-dotenv==1.2.2
PyMySQL==1.1.2
